add analytics query service
Adds a new service for querying the analytics API. Refactors the existing route controller to use the new service.
