+++ src/Controller/Controller.php
@@ -2,17 +2,8 @@
 
 namespace Drupal\asu_item_analytics\Controller;
 
-use Drupal\Core\Config\ConfigFactoryInterface;
+use Drupal\asu_item_analytics\Service\AnalyticsQueryService;
 use Drupal\Core\Controller\ControllerBase;
-use Google\Analytics\Data\V1beta\Client\BetaAnalyticsDataClient;
-use Google\Analytics\Data\V1beta\DateRange;
-use Google\Analytics\Data\V1beta\Dimension;
-use Google\Analytics\Data\V1beta\Filter;
-use Google\Analytics\Data\V1beta\Filter\InListFilter;
-use Google\Analytics\Data\V1beta\FilterExpression;
-use Google\Analytics\Data\V1beta\FilterExpressionList;
-use Google\Analytics\Data\V1beta\Metric;
-use Google\Analytics\Data\V1beta\RunReportRequest;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
 
@@ -22,20 +13,20 @@ use Symfony\Component\HttpFoundation\JsonResponse;
 class Controller extends ControllerBase {
 
   /**
-   * The configuration service.
+   * The analytics query service.
    *
-   * @var \Drupal\Core\Config\Config
+   * @var \Drupal\asu_item_analytics\Service\AnalyticsQueryService
    */
-  protected $config;
+  protected $analyticsQueryService;
 
   /**
-   * Constructs a new Controller object.
+   * Constructs a new AnalyticsController object.
    *
-   * @param \Drupal\Core\Config\ConfigFactoryInterface $config_factory
-   *   The configuration factory.
+   * @param \Drupal\asu_item_analytics\Service\AnalyticsQueryService $analytics_query_service
+   *   The analytics query service.
    */
-  public function __construct(ConfigFactoryInterface $config_factory) {
-    $this->config = $config_factory->get('asu_item_analytics.settings');
+  public function __construct(AnalyticsQueryService $analytics_query_service) {
+    $this->analyticsQueryService = $analytics_query_service;
   }
 
   /**
@@ -43,7 +34,7 @@ class Controller extends ControllerBase {
    */
   public static function create(ContainerInterface $container) {
     return new static(
-      $container->get('config.factory')
+      $container->get('asu_item_analytics.query')
     );
   }
 
@@ -57,81 +48,9 @@ class Controller extends ControllerBase {
    *   The JSON response.
    */
   public function monthly($node) {
-
     $path = $node->toUrl()->toString();
-
-    $credentialsPath = $this->config->get('credentials_path');
-    $client = new BetaAnalyticsDataClient(['credentials' => $credentialsPath]);
-    $propertyId = $this->config->get('property_id');
-    $eventName = $this->config->get('event_name');
-
-    $request = (new RunReportRequest())
-      ->setProperty('properties/' . $propertyId)
-      ->setDateRanges(
-        [
-          new DateRange(
-            [
-              // We started collecting data for the event in question in 2024.
-              'start_date' => '2024-01-01',
-              'end_date' => 'today',
-            ]
-          ),
-        ]
-    )
-      ->setDimensions(
-        [new Dimension(
-            [
-              'name' => 'yearMonth',
-            ]
-          ),
-        ]
-    )
-      ->setDimensionFilter(
-        new FilterExpression(
-            [
-              'and_group' => new FilterExpressionList(
-                [
-                  'expressions' => [
-                    new FilterExpression(
-                    [
-                      'filter' => new Filter(
-                        [
-                          'field_name' => 'eventName',
-                          'in_list_filter' => new InListFilter(['values' => [$eventName]]),
-                        ]
-                      ),
-                    ]
-                    ),
-                    new FilterExpression(
-                        [
-                          'filter' => new Filter(
-                            [
-                              'field_name' => 'pagePath',
-                              'in_list_filter' => new InListFilter(['values' => [$path]]),
-                            ]
-                          ),
-                        ]
-                    ),
-                  ],
-                ]
-              ),
-            ]
-        )
-    )
-      ->setMetrics(
-        [new Metric(
-            [
-              'name' => 'eventCount',
-            ]
-          ),
-        ]
-    );
-    $response = $client->runReport($request);
-
-    $data = [];
-    foreach ($response->getRows() as $row) {
-      $data[$row->getDimensionValues()[0]->getValue()] = $row->getMetricValues()[0]->getValue();
-    }
+    $aq = \Drupal::service('asu_item_analytics.query');
+    $data = $aq->nodeMonthly($node);
     return new JsonResponse($data);
   }
